Remove deprecated character actions and shared functionality
- **pkg/character/character_actions.go**:
  - Deleted file, removing outdated spellcasting, attack, and HP modification logic.

- **pkg/shared/damage_breaker.go**:
  - Removed file, eliminating unused `DamageBreaker` type and related constants.

- **pkg/core/martial_attacks/martial_attacks.go**:
  - Deleted file, removing deprecated martial attack logic.

- **pkg/shared/saving_throws.go**:
  - Removed file, including outdated saving throw calculations and logic.

- **pkg/core/entity.go**:
  - Refactored `Entity` interface to remove legacy methods for HP management and saving throws.
  - Adjusted method signatures for consistency with refactored packages.

- **pkg/core/equipment_manager/equipment_manager.go**:
  - Introduced new package to manage entity equipment, including armor and weapons.
  - Added methods for weapon proficiency and retrieving or setting equipment details.

+++ pkg/core/util.go
@@ -21,7 +21,7 @@ var monsterPBTable = []ChallengeRatingPB{
 	{[]float64{29, 30}, 9},
 }
 
-var characterPBTable = map[int]int{
+var characterPBTable = map[uint8]int{
 	1: 2, 2: 2, 3: 2, 4: 2,
 	5: 3, 6: 3, 7: 3, 8: 3,
 	9: 4, 10: 4, 11: 4, 12: 4,
@@ -29,7 +29,7 @@ var characterPBTable = map[int]int{
 	17: 6, 18: 6, 19: 6, 20: 6,
 }
 
-func GetCharacterProficiencyBonus(level int) (int, error) {
+func GetCharacterProficiencyBonus(level uint8) (int, error) {
 	if bonus, exists := characterPBTable[level]; exists {
 		return bonus, nil
 	}
